fin debug & secu user

+++ src/src/Controller/UserController.php
@@ -10,11 +10,21 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
+/*
+ * Controller pour l'utilisateur
+ * Un autre controleur pour la gestion des utilisateurs par les admins
+ * c'est plus de clarté
+ */
 
+/**
+ * Class UserController
+ * @package App\Controller
+ * @Route("/user")
+ */
 class UserController extends AbstractController
 {
     /**
-     * @Route("/user", name="user_index", methods={"GET"})
+     * @Route("/", name="user_index", methods={"GET"})
      */
     public function index(UserRepository $userRepository): Response
     {
@@ -25,17 +35,18 @@ class UserController extends AbstractController
 
     /**
      *
-     * @Route("/user/{id}", name="user_show", methods={"GET"})
+     * @Route("/{id}", name="user_show", methods={"GET"})
      */
     public function show(User $user): Response
     {
+        //TODO: Voir pour que le LEADER de la Team puisse reste le mdp avec en mail
         return $this->render('user/show.html.twig', [
             'user' => $user,
         ]);
     }
 
     /**
-     * @Route("/user/{id}/edit", name="user_edit", methods={"GET","POST"})
+     * @Route("/{id}/edit", name="user_edit", methods={"GET","POST"})
      */
     public function edit(Request $request, User $user): Response
     {
@@ -55,23 +66,29 @@ class UserController extends AbstractController
                 'form' => $form->createView(),
             ]);
         }else{
-            //TODO: Faire un message d'erreur
+            //TODO: Faire un message d'erreur sur l'édite
             return $this->redirectToRoute('home');
         }
 
     }
 
     /**
-     * @Route("/user/{id}", name="user_delete", methods={"DELETE"})
+     * @Route("/{id}", name="user_delete", methods={"DELETE"})
      */
     public function delete(Request $request, User $user): Response
     {
-        if ($this->isCsrfTokenValid('delete' . $user->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($user);
-            $entityManager->flush();
+        if($user = $this->$user || is_granted('LEADER') || is_granted('ADMINISTRATOR')){
+            if ($this->isCsrfTokenValid('delete' . $user->getId(), $request->request->get('_token'))) {
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->remove($user);
+                $entityManager->flush();
+            }
+
+            return $this->redirectToRoute('user_index');
+        }else{
+            //TODO: Faire un message d'erreur sur le delete
+            return $this->redirectToRoute('home');
         }
 
-        return $this->redirectToRoute('user_index');
     }
 }
